requirements.txt
pillow
textwrap